Adds day 9, breaks day 7. Take that, day 7!

+++ 2019/day07/day07_test.go
@@ -9,7 +9,7 @@ import (
 
 type testData struct {
 	input    string
-	expected int
+	expected int64
 }
 
 func TestSolvePartOne(t *testing.T) {
